+++ newpolka/tests/libFuzzer/test_poly32.c
@@ -3,9 +3,9 @@
 #include <string.h>
 #include <stdio.h>
 
-extern int LLVMFuzzerTestOneInput(const int *data, size_t dataSize) {
+extern int LLVMFuzzerTestOneInput(const long *data, size_t dataSize) {
 	unsigned int dataIndex = 0;
-	int dim;
+	long dim;
 	FILE *fp;
 	fp = fopen("out32.txt", "w+");
